requirements.txt
python-dateutil>=2.8.2
rich>=13.7.0
typer>=0.9.0
pydantic>=2.6.1
openai>=1.12.0
python-dotenv>=1.0.0
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
mypy>=1.8.0 